sistema_evaluacion_codigo/sistema_evaluacion_codigo/development.py
import environ

from sistema_evaluacion_codigo.base import *

env = environ.Env()
env_file = os.path.join(BASE_DIR, '.env')
environ.Env.read_env(env_file)

SECRET_KEY = env("SECRET_KEY_DEVELOPMENT")
DEBUG = True

# Database
# https://docs.djangoproject.com/en/2.0/ref/settings/#databases
DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.sqlite3',
        'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
    }
}
