sistema_evaluacion_codigo/requirements.txt
Django==2.0.4
django-bootstrap-datepicker-plus==3.0.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.10
django-extensions==2.1.0
Pillow==5.2.0
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
tzlocal==1.5.1
django-environ==0.4.5
mock==2.0.0
mysqlclient==1.3.12
gunicorn==19.7.1
